feat: enable 24 bit addresses
- Even the SNES had 24 bit addressing. This makes the PC register a 24 bit (well 32 bit due to rust limitations but the hardware would be techincally 24 bit). This is the only register that is >16 bit but it is OK because it is only used internally and can't interface with the data bus
- The general purpose a1-3 registers have been replaced with an high/low address registers. This is kind of a work around for the 16 bit data bus/register size. The two registers will be combined in address instructions into 24 bits

+++ toolchain/src/types/object.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 #[derive(Serialize, Deserialize, Debug)]
 pub struct SymbolDefinition {
     pub name: String,
-    pub offset: u16,
+    pub offset: u32,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
